backend/requirements.txt
Flask==2.2.5
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.1
PyJWT==2.1.0
Werkzeug==2.2.2
flask_cors